Fix measurement-deferral correctness and unblock workspace CI
Pass correctness (mlir_bridge/src/passes/measurement_deferral.rs):
- Fix the deferred-correction composition order. The controlled block must
  realise C(U·V†) applied after the unconditional V, but compose_u_v_dagger
  was emitting V†·U, which only agrees when U and V commute (all current
  tests use an identity else-branch). Emit V† first, then U, so the operator
  is U·V†. Documented the deferred-measurement identity at the function.
- Make the terminal-measure lookup panic-free (replace `output_map[&0]`
  indexing with a checked get + typed DeferError).
- Remove unused `Operation` import; collapse nested `if let` (clippy).

Unblock workspace clippy `-D warnings` (pre-existing debt, masked while
quon_core failed to compile first):
- zx/src/rewrite.rs: allow(dead_code) on the remaining stub rewrite rules.
- quon_core/src/optimization.rs: allow(clippy::implicit_saturating_sub) so
  the Flux-verified explicit branch is kept.
- frontend/src/lower.rs: drop redundant `as usize`; allow type_complexity.
- mlir_bridge clifford_t_opt / compiler_uncomputation: drop unused imports.
- mlir_bridge monadic_lower example + monadic_lowering tests: drop needless mut.

+++ frontend/src/lower.rs
@@ -191,7 +191,7 @@ impl<'c> LoweringCtx<'c> {
                 .map(|_| (qubit, self.location))
                 .collect::<Vec<_>>(),
         );
-        let mut wires: Vec<Value<'c, 'c>> = Vec::with_capacity(block.argument_count() as usize);
+        let mut wires: Vec<Value<'c, 'c>> = Vec::with_capacity(block.argument_count());
         for i in 0..block.argument_count() {
             let arg = block
                 .argument(i)
@@ -489,6 +489,7 @@ fn circuit_ref_op<'c>(
     Ok(operation)
 }
 
+#[allow(clippy::type_complexity)]
 fn collect_gate_placements(expr: &Sp<Expr>) -> Result<Vec<(Sp<Expr>, Sp<Expr>)>, LowerError> {
     match &expr.0 {
         Expr::CircuitBlock(stmts) => {
